+++ src/daomanage/datadef.go
@@ -5,13 +5,13 @@ import (
 )
 
 type IDManager struct {
-	PID int64 `bson:"pid" json:"pid"`
-	CID int64 `bson:"cid" json:"cid"`
-	RID int64 `bson:"rid" json:"rid"`
+	PID int `bson:"pid" json:"pid"`
+	CID int `bson:"cid" json:"cid"`
+	RID int `bson:"rid" json:"rid"`
 }
 
 type Problem struct {
-	PID          int64  `bson:"pid" json:"pid"`
+	PID          int    `bson:"pid" json:"pid"`
 	Title        string `bson:"title" json:"title"`
 	Description  string `bson:"description" json:"description"`
 	Time         string `bson:"time" json:"time"`
@@ -20,7 +20,7 @@ type Problem struct {
 	Output       string `bson:"output" json:"output"`
 	SimpleInput  string `bson:"simpleinput" json:"simpleinput"`
 	SimpleOutput string `bson:"simpleoutput" json:"simpleoutput"`
-	Solved       int64  `bson:"solved" json:"solved"`
+	Solved       int    `bson:"solved" json:"solved"`
 	Display      bool   `bson:"display" json:"display"`
 	Author       string `bson:"author" json:"author"`
 }
@@ -35,38 +35,38 @@ type User struct {
 }
 
 type Contest struct {
-	CID int64 `bson:"cid" json:"cid"`
+	CID int `bson:"cid" json:"cid"`
 	// pid1;pid2;pid3
 	ContestName string `bson:"contestname" json:"contestname"`
 	ProblemList string `bson:"problemlist" json:"problemlist"`
-	StartTime   int64  `bson:"starttime" json:"starttime"`
-	EndTime     int64  `bson:"endtime" json:"endtime"`
+	StartTime   int    `bson:"starttime" json:"starttime"`
+	EndTime     int    `bson:"endtime" json:"endtime"`
 }
 
 type ContestProblem struct {
-	CID    int64 `bson:"cid" json:"cid"`
-	PID    int64 `bson:"pid" json:"pid"`
-	Solved int64 `bson:"solved" json:"solved"`
-	Score  int64 `bson:"score" json:"score"`
+	CID    int `bson:"cid" json:"cid"`
+	PID    int `bson:"pid" json:"pid"`
+	Solved int `bson:"solved" json:"solved"`
+	Score  int `bson:"score" json:"score"`
 }
 
 type SubmitQueue struct {
-	PID  int64  `bson:"pid" json:"pid"`
+	PID  int    `bson:"pid" json:"pid"`
 	Code string `bson:"code" json:"code"`
 	Lang string `bson:"lang" json:"lang"`
 }
 
 type ExSubmitQueue struct {
 	ID   bson.ObjectId `bson:"_id" json:"_id"`
-	PID  int64         `bson:"pid" json:"pid"`
+	PID  int           `bson:"pid" json:"pid"`
 	Code string        `bson:"code" json:"code"`
 	Lang string        `bson:"lang" json:"lang"`
 }
 
 type RuntimeStatus struct {
 	Index   string `bson:"_index" json:"_index"`
-	RID     int64  `bson:"rid" json:"rid"`
-	PID     int64  `bson:"pid" json:"pid"`
+	RID     int    `bson:"rid" json:"rid"`
+	PID     int    `bson:"pid" json:"pid"`
 	SbmTime string `bson:"sbmtime" json:"sbmtime"`
 	Status  string `bson:"status" json:"status"`
 	Code    string `bson:"code" json:"code"`
@@ -89,8 +89,8 @@ type StatusInfoSet struct {
 }
 
 type RuntimeStatusInfo struct {
-	RID     int64  `bson:"rid" json:"rid"`
-	PID     int64  `bson:"pid" json:"pid"`
+	RID     int    `bson:"rid" json:"rid"`
+	PID     int    `bson:"pid" json:"pid"`
 	SbmTime string `bson:"sbmtime" json:"sbmtime"`
 	Status  string `bson:"status" json:"status"`
 	Memory  string `bson:"memory" json:"memory"`
@@ -106,8 +106,8 @@ type ProblemInfoSet struct {
 }
 
 type ProblemInfo struct {
-	PID         int64  `bson:"pid" json:"pid"`
+	PID         int    `bson:"pid" json:"pid"`
 	ProblemName string `bson:"title" json:"title"`
-	Solved      int64  `bson:"solved" json:"solved"`
+	Solved      int    `bson:"solved" json:"solved"`
 	Author      string `bson:"author" json:"author"`
 }
